+++ storjnode/util.py
@@ -11,6 +11,7 @@ def address_to_node_id(address):
     # e.g. address = api.get_address(wif)
     return a2b_hashed_base58(address)[1:]
 
+
 def full_path(path):
     """Resolves, sym links, rel paths, variables, and tilds to abs paths."""
     return os.path.realpath(os.path.expandvars(os.path.expanduser(path)))