+++ tests/network/__init__.py
@@ -1,6 +1,5 @@
 from . file_transfer import * # NOQA
 from . api import * # NOQA
-from . map import * # NOQA
 
 
 if __name__ == "__main__":